feat(cms): finish

+++ server/controller/weapon.js
@@ -2,42 +2,42 @@ const weaponDb = require("../model/weapon");
 const weaponService = require("../services/weapon");
 const uploadImage = require("../helpers/uploadImage");
 const deleteImage = require("../helpers/deleteImage");
-const validator = require("../helpers/validator")
+const validator = require("../helpers/validator");
 
 exports.create = async (req, res) => {
   const data = req.body;
   if (req.files && req.files.image) data.image = req.files.image;
   const validate = weaponService.validateCreate(data);
   if (validate.next) {
-    data.item_price = JSON.parse(data.item_price);
-    data.item_price = validator.cleanPriceObj(data.item_price);
-    const uploads = await uploadImage(req.files.image, "weapons");
-    if (uploads) {
-      const weapon = new weaponDb({
+    try {
+      const exists = await weaponDb.findOne({
         name: data.name,
-        image: uploads,
         item_type: data.item_type,
-        item_rarity: data.item_rarity,
-        item_price: data.item_price,
       });
-      weapon
-        .save(weapon)
-        .then((data) => {
-          res.status(200).send({
-            status: 200,
-            data,
+      if (!exists) {
+        const uploads = await uploadImage(req.files.image, "weapons");
+        if (uploads) {
+          data.item_price = JSON.parse(data.item_price);
+          data.item_price = validator.cleanPriceObj(data.item_price);
+          const weapon = new weaponDb({
+            name: data.name,
+            image: uploads,
+            item_type: data.item_type,
+            item_rarity: data.item_rarity,
+            item_price: data.item_price,
           });
-        })
-        .catch((err) => {
-          res.status(500).send({
-            status: 500,
-            error: err,
+          weapon.save(weapon).then((data) => {
+            res.status(200).send({
+              status: 200,
+              data,
+            });
           });
-        });
-    } else {
+        } else throw "Falha ao cadastrar imagens!";
+      } else throw "Já existe um weapon com esse nome e tipo";
+    } catch (err) {
       res.status(500).send({
         status: 500,
-        error: "Uploads failed!",
+        error: err,
       });
     }
   } else {
@@ -86,33 +86,41 @@ exports.update = async (req, res) => {
   const data = req.body;
   const validate = weaponService.validateUpdate(data, false);
   if (validate.next) {
-    const id = data.id;
-    delete data.id;
-    data.item_price = JSON.parse(data.item_price);
-    data.item_price = validator.cleanPriceObj(data.item_price);
-    if (req.files) {
-      const actualWeapon = await weaponDb.findById(id);
-      if (deleteImage(actualWeapon.image, "weapons")) {
-        const uploads = await uploadImage(req.files.image, "weapons");
-        if (uploads) {
-          data.image = uploads;
-        }
-      }
-    }
-    weaponDb
-      .findByIdAndUpdate(id, { ...data })
-      .then((data) => {
-        res.status(200).send({
-          status: 200,
-          data,
-        });
-      })
-      .catch((err) => {
-        res.status(500).send({
-          status: 500,
-          error: err,
-        });
+    try {
+      const id = req.params.id;
+      const exists = await weaponDb.findOne({
+        name: data.name,
+        item_type: data.item_type,
+        _id: { $ne: id },
       });
+      if (!exists) {
+        delete data.id;
+        data.item_price = JSON.parse(data.item_price);
+        data.item_price = validator.cleanPriceObj(data.item_price);
+        if (req.files) {
+          const actualWeapon = await weaponDb.findById(id);
+          if (deleteImage(actualWeapon.image, "weapons")) {
+            const uploads = await uploadImage(req.files.image, "weapons");
+            if (uploads) {
+              data.image = uploads;
+            }
+            throw "Falha ao cadastrar imagens!";
+          }
+          throw "Falha ao deletar imagens!";
+        }
+        weaponDb.findByIdAndUpdate(id, { ...data }).then((data) => {
+          res.status(200).send({
+            status: 200,
+            data,
+          });
+        });
+      } else throw "Já existe um weapon com esse nome e tipo";
+    } catch (err) {
+      res.status(500).send({
+        status: 500,
+        error: err,
+      });
+    }
   } else {
     res.status(500).send({
       status: 500,
@@ -131,10 +139,10 @@ exports.delete = async (req, res) => {
       res.status(200).send({
         status: 200,
       });
-    } catch (e) {
+    } catch (err) {
       res.status(500).send({
         status: 500,
-        error: "Houve um erro ao tentar deletar o weapon",
+        error: err,
       });
     }
   } else {
